libobs: Add "dictionary" property type and helper API
The dictionary property can be used to allow a user to define a collection
of values with associated types which can then be interpreted by a source
or other plugin. A helper API is added to allow developers to access
and manipulate dictionary entries without needing to be aware of their
internal representation.