obs-filters: Support relative multipliers in scale filters
The newly-supported syntax is a floating point number prefixed with
the character "x". So "x2" would mean double the size of the input
video; "x0.33" would mean a third of the input video's size.

The code that parses these strings is rearranged slightly to make
it easier to test three or more different formats.